Initial Commit of the ViAppleGrab Code
This initial commit has a working version of ViAppleGrab with only minor
modifications from the version of ViAppleGrab which was originally used
for the Graphics Interface studies. New additions include the creation
of a game installer and a warning dialog which prevents the user from
starting the system if the PS3 has not been properly networked and the
Move.Me server started. Also, by default, the results folder opens upon
the completion of the game.