EditUser form is now complete, UserInfo removed
The UserInfo form is no longer needed since the EditUser form is more
comprehensive and much more usable. In addition, the InitMenu form has
been laid out so that each of the steps is sequential and very
straightforward. the UserSelection form still needs to be updated to
reflect the changes in the User.xml file structure and to handle default
configuration settings for the new camp abilities study.